requirements.txt
torch>=2.0.0
torchvision>=0.15.0
captum>=0.7.0
lime>=0.2.0
grad-cam>=1.5.0
matplotlib>=3.7.0
pyyaml>=6.0.0
numpy>=1.24.0
